refactor to ObservationsContainer
